feat: Enhance obstacle challenge logging and replay functionality, complete parallel parking, and update docstrings.
- Added detailed docstrings to DriveCommand and LogRecord classes for better clarity on their purpose and usage.
- Improved PickleLogWriter with enhanced documentation and error handling.
- Implemented build_log_record function with comprehensive docstring for creating LogRecord instances.
- Updated load_log_records function to include error handling and documentation for loading LogRecord objects.
- Enhanced format_log_record and draw_replay_overlay functions with detailed docstrings for better understanding.
- Refactored replay_video function to improve documentation and error handling.
- Added new parallel parking functionality with a dedicated state machine and runner.
- Introduced color threshold tuner utility for real-time HSV threshold adjustments.
- Updated tests to cover new features and ensure compatibility with existing functionality.
- Added documentation for CLI and improved code organization.

+++ PiClient/piclient-core/src/piclient/core/vision/async_obstacle.py
@@ -41,7 +41,7 @@ class ObstacleChallengeAsyncMultiprocessingVisionProcessor(ObstacleChallengeVisi
         """
         return await self.loop.run_in_executor(self._executor, super(ObstacleChallengeAsyncMultiprocessingVisionProcessor, self).get_walls_and_obstacles, frame)
 
-    async def get_target_async(self, walls_and_obstacles: WallsAndObstacles) -> tuple[int, int] | None:
+    async def get_target_async(self, walls_and_obstacles: WallsAndObstacles) -> tuple[float, float] | None:
         """Async wrapper around the synchronous target computation method.
 
         :param walls_and_obstacles: Walls and obstacles data.
